+++ src/textual/reactive.py
@@ -32,6 +32,10 @@ if TYPE_CHECKING:
 ReactiveType = TypeVar("ReactiveType")
 
 
+class TooManyComputesError(Exception):
+    """Raised when an attribute has public and private compute methods."""
+
+
 @rich.repr.auto
 class Reactive(Generic[ReactiveType]):
     """Reactive descriptor.
@@ -85,9 +89,9 @@ class Reactive(Generic[ReactiveType]):
             # Attribute already has a value
             return
 
-        compute_method = getattr(obj, f"compute_{name}", None)
+        compute_method = getattr(obj, self.compute_name, None)
         if compute_method is not None and self._init:
-            default = getattr(obj, f"compute_{name}")()
+            default = compute_method()
         else:
             default_or_callable = self._default
             default = (
@@ -122,7 +126,12 @@ class Reactive(Generic[ReactiveType]):
 
     def __set_name__(self, owner: Type[MessageTarget], name: str) -> None:
         # Check for compute method
-        if hasattr(owner, f"compute_{name}"):
+        public_compute = f"compute_{name}"
+        private_compute = f"_compute_{name}"
+        compute_name = (
+            private_compute if hasattr(owner, private_compute) else public_compute
+        )
+        if hasattr(owner, compute_name):
             # Compute methods are stored in a list called `__computes`
             try:
                 computes = getattr(owner, "__computes")
@@ -135,7 +144,7 @@ class Reactive(Generic[ReactiveType]):
         self.name = name
         # The internal name where the attribute's value is stored
         self.internal_name = f"_reactive_{name}"
-        self.compute_name = f"compute_{name}"
+        self.compute_name = compute_name
         default = self._default
         setattr(owner, f"_default_{name}", default)
 
@@ -167,11 +176,13 @@ class Reactive(Generic[ReactiveType]):
 
         name = self.name
         current_value = getattr(obj, name)
-        # Check for validate function
-        validate_function = getattr(obj, f"validate_{name}", None)
-        # Call validate
-        if callable(validate_function):
-            value = validate_function(value)
+        # Check for private and public validate functions.
+        private_validate_function = getattr(obj, f"_validate_{name}", None)
+        if callable(private_validate_function):
+            value = private_validate_function(value)
+        public_validate_function = getattr(obj, f"validate_{name}", None)
+        if callable(public_validate_function):
+            value = public_validate_function(value)
         # If the value has changed, or this is the first time setting the value
         if current_value != value or self._always_update:
             # Store the internal value
@@ -266,7 +277,10 @@ class Reactive(Generic[ReactiveType]):
             try:
                 compute_method = getattr(obj, f"compute_{compute}")
             except AttributeError:
-                continue
+                try:
+                    compute_method = getattr(obj, f"_compute_{compute}")
+                except AttributeError:
+                    continue
             current_value = getattr(
                 obj, f"_reactive_{compute}", getattr(obj, f"_default_{compute}", None)
             )